Facilitate API interaction through wis2box/api/__init__.py
- Wrap backend and config methods in wis2box/api/__init__.py.
- Move CLI options from `api` node to `data` node
