+++ .github/workflows/tests-docker.yml
@@ -49,7 +49,7 @@ jobs:
         TEST_DATA: /data/wis2box/observations/malawi
       run: |
         python3 wis2box-ctl.py execute wis2box metadata discovery publish $DISCOVERY_METADATA
-        python3 wis2box-ctl.py execute wis2box api add-collection -th $TOPIC_HIERARCHY $DISCOVERY_METADATA
+        python3 wis2box-ctl.py execute wis2box data add-collection -th $TOPIC_HIERARCHY $DISCOVERY_METADATA
         python3 wis2box-ctl.py execute wis2box data ingest -th $TOPIC_HIERARCHY -p $TEST_DATA
     - name: add Italy data 🇮🇹
       env:
@@ -58,7 +58,7 @@ jobs:
         TEST_DATA: /data/wis2box/observations/italy
       run: |
         python3 wis2box-ctl.py execute wis2box metadata discovery publish $DISCOVERY_METADATA
-        python3 wis2box-ctl.py execute wis2box api add-collection -th $TOPIC_HIERARCHY $DISCOVERY_METADATA
+        python3 wis2box-ctl.py execute wis2box data add-collection -th $TOPIC_HIERARCHY $DISCOVERY_METADATA
         python3 wis2box-ctl.py execute wis2box data ingest -th $TOPIC_HIERARCHY -p $TEST_DATA
     - name: add Algeria data 🇩🇿
       env:
@@ -67,7 +67,7 @@ jobs:
         TEST_DATA: /data/wis2box/observations/algeria
       run: |
         python3 wis2box-ctl.py execute wis2box metadata discovery publish $DISCOVERY_METADATA
-        python3 wis2box-ctl.py execute wis2box api add-collection -th $TOPIC_HIERARCHY $DISCOVERY_METADATA
+        python3 wis2box-ctl.py execute wis2box data add-collection -th $TOPIC_HIERARCHY $DISCOVERY_METADATA
         python3 wis2box-ctl.py execute wis2box data ingest -th $TOPIC_HIERARCHY -p $TEST_DATA
     - name: sync stations 🔄
       run: |
