+++ wis2box/__init__.py
@@ -27,7 +27,6 @@ from wis2box.data import data
 from wis2box.env import environment
 from wis2box.metadata import metadata
 from wis2box.pubsub import pubsub
-from wis2box.api import api
 from wis2box.auth import auth
 
 
@@ -41,6 +40,5 @@ def cli():
 cli.add_command(environment)
 cli.add_command(data)
 cli.add_command(metadata)
-cli.add_command(api)
 cli.add_command(auth)
 cli.add_command(pubsub)
