+++ wis2box/data/__init__.py
@@ -25,16 +25,16 @@ import logging
 import click
 
 from wis2box import cli_helpers
-from wis2box.api.backend import load_backend
+from wis2box.api import (setup_collection, remove_collection,
+                         delete_collections_by_retention)
 from wis2box.env import (STORAGE_SOURCE, STORAGE_ARCHIVE, STORAGE_PUBLIC,
                          STORAGE_DATA_RETENTION_DAYS, STORAGE_INCOMING)
 from wis2box.handler import Handler
-from wis2box.storage import put_data
+from wis2box.metadata.discovery import DiscoveryMetadata
+from wis2box.storage import put_data, move_data, list_content, delete_data
 from wis2box.util import older_than, walk_path
+from wis2box.topic_hierarchy import validate_and_load
 
-from wis2box.storage import move_data
-from wis2box.storage import list_content
-from wis2box.storage import delete_data
 
 LOGGER = logging.getLogger(__name__)
 
@@ -86,8 +86,49 @@ def clean_data(source_path: str, days: int) -> None:
             LOGGER.debug(f"{object['basedir']} less than {days} days old")
 
     LOGGER.debug('Cleaning API indexes')
-    backend = load_backend()
-    backend.delete_collections_by_retention(days)
+    delete_collections_by_retention(days)
+
+
+def generate_collection_metadata(mcf: dict) -> dict:
+    """
+    Generate collection metadata from metadata control file
+
+    :param mcf: `dict` of MCF file
+
+    :returns: `dict` of API collection metadata
+    """
+
+    LOGGER.debug('Parsing discovery metadata')
+
+    dm = DiscoveryMetadata()
+    record = dm.parse_record(mcf)
+    generated = dm.generate(record)
+
+    LOGGER.debug('Creating collection configuration')
+
+    bbox = [
+        generated['geometry']['coordinates'][0][0][0],
+        generated['geometry']['coordinates'][0][0][1],
+        generated['geometry']['coordinates'][0][2][0],
+        generated['geometry']['coordinates'][0][2][1],
+    ]
+
+    kw = record['identification']['keywords']
+
+    keywords = set([k for k in kw.values() for k in kw])
+
+    return {
+        'id': generated['id'],
+        'type': 'feature',
+        'title': generated['properties']['title'],
+        'description': generated['properties']['description'],
+        'keywords': list(keywords),
+        'bbox': bbox,
+        'links': generated['links'],
+        'id_field': 'id',
+        'time_field': 'resultTime',
+        'title_field': 'id',
+    }
 
 
 @click.group()
@@ -100,7 +141,7 @@ def data():
 @click.pass_context
 @cli_helpers.OPTION_VERBOSITY
 def archive(ctx, verbosity):
-    """Move data from incoming storage to archive storage """
+    """Move data from incoming storage to archive storage"""
 
     source_path = f'{STORAGE_SOURCE}/{STORAGE_INCOMING}'
     archive_path = f'{STORAGE_SOURCE}/{STORAGE_ARCHIVE}'
@@ -151,6 +192,66 @@ def ingest(ctx, topic_hierarchy, path, recursive, verbosity):
     click.echo("Done")
 
 
+@click.command()
+@click.pass_context
+@cli_helpers.ARGUMENT_FILEPATH
+@cli_helpers.OPTION_TOPIC_HIERARCHY
+@cli_helpers.OPTION_VERBOSITY
+def add_collection(ctx, filepath, topic_hierarchy, verbosity):
+    """Add collection index to API backend"""
+
+    if topic_hierarchy is None:
+        raise click.ClickException('Missing -th/--topic-hierarchy')
+
+    click.echo('Generating collection metadata')
+    meta = generate_collection_metadata(filepath.read())
+
+    th, _ = validate_and_load(topic_hierarchy)
+    click.echo(f'Adding collection: {th.dotpath}')
+    setup_collection(th.dotpath, meta=meta)
+
+    click.echo("Done")
+
+
+@click.command()
+@click.pass_context
+@cli_helpers.OPTION_TOPIC_HIERARCHY
+@cli_helpers.OPTION_VERBOSITY
+def delete_collection(ctx, topic_hierarchy, verbosity):
+    """Delete collection from api backend"""
+
+    if topic_hierarchy is None:
+        raise click.ClickException('Missing -th/--topic-hierarchy')
+
+    click.echo(f'Deleting collection: {topic_hierarchy}')
+
+    th, _ = validate_and_load(topic_hierarchy)
+    remove_collection(th.dotpath)
+
+    click.echo('Done')
+
+
+@click.command()
+@click.pass_context
+@cli_helpers.OPTION_TOPIC_HIERARCHY
+@cli_helpers.OPTION_PATH
+@cli_helpers.OPTION_RECURSIVE
+@cli_helpers.OPTION_VERBOSITY
+def add_collection_items(ctx, topic_hierarchy, path, recursive, verbosity):
+    """Add collection items to API backend"""
+
+    click.echo(f'Adding GeoJSON files to collection: {topic_hierarchy}')
+    for file_to_process in walk_path(path, '.*.geojson$', recursive):
+        click.echo(f'Adding {file_to_process}')
+        handler = Handler(file_to_process, topic_hierarchy)
+        handler.publish()
+
+    click.echo('Done')
+
+
 data.add_command(archive)
 data.add_command(clean)
 data.add_command(ingest)
+data.add_command(add_collection)
+data.add_command(delete_collection)
+data.add_command(add_collection_items)
