+++ wis2box/metadata/discovery.py
@@ -26,8 +26,8 @@ import logging
 from pygeometa.schemas.ogcapi_records import OGCAPIRecordOutputSchema
 
 from wis2box import cli_helpers
-from wis2box.api.backend import load_backend
-from wis2box.api.config import load_config
+from wis2box.api import (setup_collection, upsert_collection_item,
+                         delete_collection_item)
 from wis2box.env import API_URL, BROKER_PUBLIC
 from wis2box.metadata.base import BaseMetadata
 from wis2box.util import remove_auth_from_url
@@ -121,9 +121,7 @@ def publish_collection() -> bool:
         'title_field': 'title',
     }
 
-    api_config = load_config()
-    collection = api_config.prepare_collection(meta)
-    api_config.add_collection(meta['id'], collection)
+    setup_collection('discovery-metadata', meta=meta)
 
     return True
 
@@ -146,8 +144,7 @@ def publish(ctx, filepath, verbosity):
         dm = DiscoveryMetadata()
         record = dm.parse_record(filepath.read())
         record = dm.generate(record)
-        backend = load_backend()
-        backend.upsert_collection_items('discovery-metadata', [record])
+        upsert_collection_item('discovery-metadata', record)
         publish_collection()
     except Exception as err:
         raise click.ClickException(err)
@@ -163,8 +160,7 @@ def unpublish(ctx, identifier, verbosity):
     """Deletes a discovery metadata record from the catalogue"""
 
     click.echo('Unpublishing discovery metadata {identifier}')
-    backend = load_backend()
-    backend.delete_collection_item('discovery-metadata', identifier)
+    delete_collection_item('discovery-metadata', identifier)
 
 
 discovery_metadata.add_command(publish)
